Update Owl_C.lua
Code refactoring.

Added the ability to have more than one button showing.

Added the ability to have the buttons show for any period of time required.

Make the resource try to render ONLY when there's something to render, this saves CPU time.

Deleted `the` from `to acces the`, since this should be added by passing the name.

Added comments to explain the changes.